TYRUS-248: Removed o.g.t.w.Extension, using TyrusExtension instead.
+ related code cleanup
