+++ core/src/main/java/org/glassfish/tyrus/core/RemoteEndpointWrapper.java
@@ -54,7 +54,7 @@ import javax.websocket.SendHandler;
 import javax.websocket.SendResult;
 
 import org.glassfish.tyrus.spi.RemoteEndpoint;
-import static org.glassfish.tyrus.core.Utils.checkNotNull;
+import static org.glassfish.tyrus.websockets.Utils.checkNotNull;
 
 /**
  * Wraps the {@link javax.websocket.RemoteEndpoint} and represents the other side of the websocket connection.
