+++ core/src/test/java/org/glassfish/tyrus/core/RemoteEndpointWrapperTest.java
@@ -452,7 +452,7 @@ public class RemoteEndpointWrapperTest {
 
         @Override
         public Set<Extension> getInstalledExtensions() {
-            return null;
+            return Collections.emptySet();
         }
 
         @Override
