+++ protocol/websocket/src/main/java/org/glassfish/tyrus/websockets/Handshake.java
@@ -46,7 +46,8 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
-import java.util.logging.Logger;
+
+import javax.websocket.Extension;
 
 import org.glassfish.tyrus.spi.UpgradeRequest;
 import org.glassfish.tyrus.spi.UpgradeResponse;
@@ -57,14 +58,11 @@ import org.glassfish.tyrus.spi.UpgradeResponse;
  */
 public final class Handshake {
 
-    private static final Logger LOGGER = Logger.getLogger(Handshake.class.getName());
-    private final List<String> enabledExtensions = Collections.emptyList();
     private boolean secure;
     private String origin;
     private String serverHostName;
     private int port = 80;
     private String resourcePath;
-    //private final Map<String, String[]> queryParams = new TreeMap<String, String[]>();
     private List<String> subProtocols = new ArrayList<String>();
     private List<Extension> extensions = new ArrayList<Extension>(); // client extensions
     // client side request!
@@ -135,7 +133,7 @@ public final class Handshake {
 
         List<String> value = request.getHeaders().get(UpgradeRequest.SEC_WEBSOCKET_EXTENSIONS);
         if (value != null) {
-            handshake.extensions = Handshake.fromHeaders(value);
+            handshake.extensions = TyrusExtension.fromHeaders(value);
         }
         handshake.secKey = SecKey.generateServerKey(new SecKey(request.getFirstHeaderValue(UpgradeRequest.SEC_WEBSOCKET_KEY)));
 
@@ -171,220 +169,6 @@ public final class Handshake {
             handshake.serverHostName = header.substring(0, i);
             handshake.port = Integer.valueOf(header.substring(i + 1));
         }
-    }
-
-    /**
-     * Parse {@link Extension} from headers (represented as {@link List} of strings).
-     *
-     * @param extensionHeaders Http Extension headers.
-     * @return list of parsed {@link Extension Extensions}.
-     */
-    private static List<Extension> fromHeaders(List<String> extensionHeaders) {
-        List<Extension> extensions = new ArrayList<Extension>();
-
-        for (String singleHeader : extensionHeaders) {
-            if (singleHeader == null) {
-                break;
-            }
-            final char[] chars = singleHeader.toCharArray();
-            int i = 0;
-            ParserState next = ParserState.NAME;
-            StringBuilder name = new StringBuilder();
-            StringBuilder paramName = new StringBuilder();
-            StringBuilder paramValue = new StringBuilder();
-            List<Extension.Parameter> params = new ArrayList<Extension.Parameter>();
-
-            do {
-                switch (next) {
-                    case NAME:
-
-                        switch (chars[i]) {
-                            case ';':
-                                next = ParserState.PARAM_NAME;
-                                break;
-                            case ',':
-                                next = ParserState.NAME;
-                                if (name.length() > 0) {
-                                    final Extension extension = new Extension(name.toString().trim());
-                                    extension.getParameters().addAll(params);
-                                    extensions.add(extension);
-                                    name = new StringBuilder();
-                                    paramName = new StringBuilder();
-                                    paramValue = new StringBuilder();
-                                    params.clear();
-                                }
-
-                                break;
-                            case '=':
-                                next = ParserState.ERROR;
-                                break;
-                            default:
-                                name.append(chars[i]);
-                        }
-
-                        break;
-
-                    case PARAM_NAME:
-
-                        switch (chars[i]) {
-                            case ';':
-                                next = ParserState.ERROR;
-                                break;
-                            case '=':
-                                next = ParserState.PARAM_VALUE;
-                                break;
-                            default:
-                                paramName.append(chars[i]);
-                        }
-
-                        break;
-
-                    case PARAM_VALUE:
-
-                        switch (chars[i]) {
-                            case '"':
-                                if (paramValue.length() > 0) {
-                                    next = ParserState.ERROR;
-                                } else {
-                                    next = ParserState.PARAM_VALUE_QUOTED;
-                                }
-                                break;
-                            case ';':
-                                next = ParserState.PARAM_NAME;
-                                params.add(new Extension.Parameter(paramName.toString().trim(), paramValue.toString().trim()));
-                                paramName = new StringBuilder();
-                                paramValue = new StringBuilder();
-                                break;
-                            case ',':
-                                next = ParserState.NAME;
-                                params.add(new Extension.Parameter(paramName.toString().trim(), paramValue.toString().trim()));
-                                paramName = new StringBuilder();
-                                paramValue = new StringBuilder();
-                                if (name.length() > 0) {
-                                    final Extension extension = new Extension(name.toString().trim());
-                                    extension.getParameters().addAll(params);
-                                    extensions.add(extension);
-                                    name = new StringBuilder();
-                                    paramName = new StringBuilder();
-                                    paramValue = new StringBuilder();
-                                    params.clear();
-                                }
-
-                                break;
-                            case '=':
-                                next = ParserState.ERROR;
-                                break;
-                            default:
-                                paramValue.append(chars[i]);
-                        }
-
-                        break;
-
-                    case PARAM_VALUE_QUOTED:
-
-                        switch (chars[i]) {
-                            case '"':
-                                next = ParserState.PARAM_VALUE_QUOTED_POST;
-                                params.add(new Extension.Parameter(paramName.toString().trim(), paramValue.toString()));
-                                paramName = new StringBuilder();
-                                paramValue = new StringBuilder();
-                                break;
-                            case '\\':
-                                next = ParserState.PARAM_VALUE_QUOTED_QP;
-                                break;
-                            case '=':
-                                next = ParserState.ERROR;
-                                break;
-                            default:
-                                paramValue.append(chars[i]);
-                        }
-
-                        break;
-
-                    case PARAM_VALUE_QUOTED_QP:
-
-                        next = ParserState.PARAM_VALUE_QUOTED;
-                        paramValue.append(chars[i]);
-                        break;
-
-                    case PARAM_VALUE_QUOTED_POST:
-
-                        switch (chars[i]) {
-                            case ',':
-                                next = ParserState.NAME;
-                                if (name.length() > 0) {
-                                    final Extension extension = new Extension(name.toString().trim());
-                                    extension.getParameters().addAll(params);
-                                    extensions.add(extension);
-                                    name = new StringBuilder();
-                                    paramName = new StringBuilder();
-                                    paramValue = new StringBuilder();
-                                    params.clear();
-                                }
-
-                                break;
-                            case ';':
-                                next = ParserState.PARAM_NAME;
-                                break;
-                            default:
-                                next = ParserState.ERROR;
-                                break;
-                        }
-
-                        break;
-
-                    // defensive error handling - just skip this one and try to parse rest.
-                    case ERROR:
-                        LOGGER.fine(String.format("Error during parsing Extension: %s", name));
-
-                        if (name.length() > 0) {
-                            name = new StringBuilder();
-                            paramName = new StringBuilder();
-                            paramValue = new StringBuilder();
-                            params.clear();
-                        }
-
-                        switch (chars[i]) {
-                            case ',':
-                                next = ParserState.NAME;
-                                if (name.length() > 0) {
-                                    final Extension extension = new Extension(name.toString().trim());
-                                    extension.getParameters().addAll(params);
-                                    extensions.add(extension);
-                                    name = new StringBuilder();
-                                    paramName = new StringBuilder();
-                                    paramValue = new StringBuilder();
-                                    params.clear();
-                                }
-
-                                break;
-                            case ';':
-                                next = ParserState.PARAM_NAME;
-                                break;
-                            default:
-                                break;
-                        }
-
-                        break;
-                }
-
-                i++;
-            } while (i < chars.length);
-
-            if ((name.length() > 0) && (next != ParserState.ERROR)) {
-                if (paramName.length() > 0) {
-                    params.add(new Extension.Parameter(paramName.toString().trim(), paramValue.toString()));
-                }
-                final Extension extension = new Extension(name.toString().trim());
-                extension.getParameters().addAll(params);
-                extensions.add(extension);
-                params.clear();
-            } else {
-                LOGGER.fine(String.format("Unable to parse Extension: %s", name));
-            }
-        }
-
-        return extensions;
     }
 
     private static StringBuilder appendPort(StringBuilder builder, int port, boolean secure) {
@@ -521,9 +305,6 @@ public final class Handshake {
         response.getHeaders().put(UpgradeRequest.CONNECTION, Arrays.asList(UpgradeRequest.UPGRADE));
         response.setReasonPhrase(UpgradeRequest.RESPONSE_CODE_MESSAGE);
         response.getHeaders().put(UpgradeResponse.SEC_WEBSOCKET_ACCEPT, Arrays.asList(secKey.getSecKey()));
-        if (!getEnabledExtensions().isEmpty()) {
-            response.getHeaders().put(UpgradeRequest.SEC_WEBSOCKET_EXTENSIONS, getSubProtocols());
-        }
 
         if (subProtocols != null && !subProtocols.isEmpty()) {
             List<String> appProtocols = application.getSupportedProtocols(subProtocols);
@@ -531,32 +312,14 @@ public final class Handshake {
                 response.getHeaders().put(UpgradeRequest.SEC_WEBSOCKET_PROTOCOL, getStringList(appProtocols));
             }
         }
-        if (!application.getSupportedExtensions().isEmpty() && !getExtensions().isEmpty()) {
-            List<Extension> intersection =
-                    intersection(getExtensions(),
-                            application.getSupportedExtensions());
-            if (!intersection.isEmpty()) {
-                application.onExtensionNegotiation(intersection);
-                response.getHeaders().put(UpgradeRequest.SEC_WEBSOCKET_EXTENSIONS, getStringList(intersection));
-            }
+
+        if (!application.getSupportedExtensions().isEmpty()) {
+            response.getHeaders().put(UpgradeRequest.SEC_WEBSOCKET_EXTENSIONS, getStringList(application.getSupportedExtensions()));
         }
 
         application.onHandShakeResponse(incomingRequest, response);
 
         writer.write(response);
-    }
-
-    List<Extension> intersection(List<Extension> requested, List<Extension> supported) {
-        List<Extension> intersection = new ArrayList<Extension>(supported.size());
-        for (Extension e : requested) {
-            for (Extension s : supported) {
-                if (e.getName().equals(s.getName())) {
-                    intersection.add(e);
-                    break;
-                }
-            }
-        }
-        return intersection;
     }
 
     public UpgradeRequest initiate(/*FilterChainContext ctx*/) {
@@ -582,23 +345,8 @@ public final class Handshake {
         this.responseListener = responseListener;
     }
 
-    List<String> getEnabledExtensions() {
-        return enabledExtensions;
-    }
-
     int getVersion() {
         return 13;
-    }
-
-    private enum ParserState {
-        NAME,
-        PARAM_NAME,
-        PARAM_VALUE,
-        PARAM_VALUE_QUOTED,
-        PARAM_VALUE_QUOTED_POST,
-        // quoted-pair - '\' escaped character
-        PARAM_VALUE_QUOTED_QP,
-        ERROR
     }
 
     /**
