+++ protocol/websocket/src/main/java/org/glassfish/tyrus/websockets/Utils.java
@@ -37,7 +37,7 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package org.glassfish.tyrus.core;
+package org.glassfish.tyrus.websockets;
 
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
