+++ tests/e2e/standard-config/src/test/java/org/glassfish/tyrus/test/e2e/ExtensionsTest.java
@@ -57,10 +57,10 @@ import javax.websocket.Session;
 import javax.websocket.server.ServerEndpoint;
 import javax.websocket.server.ServerEndpointConfig;
 
-import org.glassfish.tyrus.test.tools.TestContainer;
 import org.glassfish.tyrus.client.ClientManager;
-import org.glassfish.tyrus.core.TyrusExtension;
 import org.glassfish.tyrus.server.Server;
+import org.glassfish.tyrus.test.tools.TestContainer;
+import org.glassfish.tyrus.websockets.TyrusExtension;
 
 import org.junit.Test;
 import static org.junit.Assert.assertEquals;
